--- a/include/gmock-win32.h
+++ b/include/gmock-win32.h
@@ -54,16 +54,21 @@ struct mock_module_##func \
 #define MOCK_MODULE_FUNC0_STDCALL_CONV(m, ...) MOCK_MODULE_FUNC0_CALLCONV(__stdcall, m, __VA_ARGS__)
 #define MOCK_MODULE_FUNC0_CDECL_CONV(m, ...) MOCK_MODULE_FUNC0_CALLCONV(__cdecl, m, __VA_ARGS__)
 
+template <typename TFunc, typename TStub>
+static void patchModuleFunc_(void* mock_module_func_oldFn, TFunc func, TStub stub) { 
+	if (!mock_module_func_oldFn) 
+		mockModule_patchModuleFunc( 
+			func 
+			, reinterpret_cast< void* >( stub ) 
+			, &mock_module_func_oldFn);
+}
+
 #define MOCK_MODULE_FUNC1_(tn, constness, ct, func, ...) \
 _Pragma( "optimize( \"\", off )" ) \
 struct mock_module_##func \
 { \
-    static void hideCodeFromOptimizer() { \
-		if (!mock_module_##func::oldFn_) \
-			mockModule_patchModuleFunc( \
-				  &::func \
-				, reinterpret_cast< void* >( &mock_module_##func::stub ) \
-				, &mock_module_##func::oldFn_); \
+    static void patchModuleFunc() { \
+		::patchModuleFunc_( mock_module_##func::oldFn_, &::func, mock_module_##func::stub ); \
 	} \
     GMOCK_RESULT_(tn, __VA_ARGS__) ct func( \
         GMOCK_ARG_(tn, 1, __VA_ARGS__) gmock_a1) constness \
@@ -824,7 +829,7 @@ void mockModule_patchModuleFunc   (void*, void*, void**);
 void mockModule_restoreModuleFunc (void*, void*, void**);
 
 #define EXPECT_MODULE_FUNC_CALL(func, ...) \
-    mock_module_##func::hideCodeFromOptimizer( ); \
+	mock_module_##func::patchModuleFunc( ); \
     EXPECT_CALL(mock_module_##func::instance(), func(__VA_ARGS__))
 
 #define ON_MODULE_FUNC_CALL(func, ...) \