@Alex0vSky 2023-09-28
Using WinApi "RegCloseKey" as an example:
...