@Alex0vSky 2023-10-02
Good news.
Both `__pragma(optimize("", on))` + `__pragma(optimize("", off))` work with MSVC 2019 and its clang.
We dont need additional marcos.